added positions and serial numbers to the declarations of a block
Introduced Declaration datatype.
Added a counter serialCounter to the FState which is increased for
each declaration. The positions recorded for each declared variable
point to the first textual occurence of that variable within its
scope. In particular: Pretypings filled in by the parser do not
point to the actual declaration (there is none) but to the first use
of the variable.