moved instructions from the reasoner state to the verification state
this change has the following effect: instructions are now only effective in the
branch of the text tree that they are made in.They are automatically reset to their
previous state upon leaving the branch. In particular: For an instruction to
be global it must be made on the top level.