Added local checking.
The verification loop now produces two text trees: One that is
used for verification with macros replaced by the actual
formulas (which are only known at runtime) and one that reflects
verification progress, that consists of the original, unaltered
formulas with TextChecked annotations.

We always keep an IORef with the latest grammatically correct
text. That text is then compared with a newly parsed text and
the text to be checked is computed. After verification we
overwrite the IORef with the newly produced structure tree.

We use the TextChecked annotations to explicitly set [prove off]
options during the verification as well as verbosity settings.
The text is therefore traversed as normal, but proof tasks are
neither executed nor printed.

At this point, ontological checks remain enabled during the whole
verification. This will be changed in a future commit.