Re-implemented ontological reduction.
The implementation now is much closer to the algorithm
description which increases clarity. Furthermore, the
preparatory normalization has been clarified in the form
of SAD.Prove.Normalize.ontoPrep. Due to difference in
preparation and formula presentation, results differ
from the results of the  old implementation.

The concept of type-likes has been abandoned as their
high maintanance cost and the decrease in code modularity
could no longer be justified.