Prepare tokenizer for latex mode
Added a parameter to the tokenizer to describe the comment syntax,
since one can easily imagine that a latex document doesn't want
to be parsed with '#' as comment syntax. One the other hand, if we
still want to parse .ftl files we do still want to use the '#'
comment syntax for backwards compatibility.

The MathMode switches in the tokenizer have been removed. The
reason is that I think math mode should not be implemented by
a math mode switch in the tokenizer, but rather should be
implemented inside the parser. There just seems to be no
reason to avoid using a proper parser for parsing the dollar
signs and instead relying on primitive Text-based parsing in
the tokenizer.

Of course, this change may be a bit more contraversial and
the effect on the overall codebase should be negligeable.