Get a protype for latex parsing going
For now, which parser is used must be manually adjusted by changing
the 'parser' variable in 'Reader.hs'. This is temporary.

This latex parser is more of a proof of concept than a finished
product. One needs to carefully decide how exactly one wants to
label the different environments, and maybe whether pretypes and
macros should be inside the same type of environment. Moreover,
it might be useful to allow multiple definitions/axioms/macros/
pretypes inside the same environment. Because opening and closing
so many environments gets annoying.

A must-have feature that is not yet implemented, is being able to
label environments with '[label]'.

But now that the basic logic for latex parsing exists, all these
features should be very easy to implement, one only needs to
decide what exactly one wants.