Add tex command expansion in tokenizer
Symbols like `|`, `->`, `/\` and other logical operators and
symbols used for notation are built into the parser and can't be
easily changed. They are very unlike the macros that are directly
expanded to terms in `initFS` in `ForTheL/Base.hs`.

In order to let the parser better abstract over whether tex was
used or not, we simply normalize corresponding tex commands in
the tokenizer to the form that they would have in `.ftl`
format. So I have added tex command expansion to the tokenizer.

There is one corner case, `\in`, where it both functions as a
symbol that is used in set notation and directly parsed by the
parser and it is also a symbol pattern that I wrote in the
previous commit. Here the idea of using a parser that does not need
to know about tex or not tex breaks down. This will be handled in
a future commit.