Handle `\in` correctly
As explained in the previous commit, `\in` doesn't really fit into
the way we are tex commands either in the tokenizer or as a builtin
symbol pattern. This is because `\in` functions as both.

The problem is that we want `\in` to be treated as a symbol, so we
can't expand it to `in`. Thus I changed the parser so `\in` is
treated as `in` in the relevant set notations.