`\` can now be used for defining variables with greek letters
This is done by renaming the greek tex commands in the tokenizer.
This makes extending this vocabulary on the fly impossible, but
I don't think there is a point in working on advanced features
such as more configurable vocabularies, since by design the series
of tex features enabling tex commands has always followed the
philosophy of getting things to work in most cases fast, without
trying to write something too polished and perfect.