Allow for spaces in `topIdentifiers` (names of environments)
This change is motivated by having more natural names for top-level blocks in LaTeX formalizations. The words are immediately concatenated, so that the identifier can still be exported to TPTP. The relaxation seems to work fine with the old ASCII syntax, since `topIdentifiers` either occur in the header of top-level blocks, where they are terminated by a dot, or in references, where they are surrounded by parentheses. With TeX syntax, they are either surrounded by brackets or parentheses. The tests still pass after this change.